requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
python-dotenv==1.0.1
jinja2==3.1.5
aiofiles==24.1.0
pydantic-settings==2.7.1
